Update and rename Part2.cs to GetMethod1.cs
